Changed typedef of iid_t to 32 bits on 64 bit machines too. (4-byte required by BDB access method) Updated printfs all around
git-svn-id: http://libpaxos.svn.sourceforge.net/svnroot/libpaxos/trunk/libpaxos2@115 00aac72d-264a-0410-add3-e1605ea6af76

+++ libpaxos.h
@@ -1,6 +1,7 @@
 #ifndef _LIBPAXOS_H_
 #define _LIBPAXOS_H_
 #include <sys/types.h>
+#include <stdint.h>
 #include "paxos_config.h"
 
 /* 
@@ -15,7 +16,7 @@
     Alias for instance identificator and ballot number.
 */
 typedef unsigned int ballot_t;
-typedef long unsigned int iid_t;
+typedef uint32_t iid_t;
 
 /* 
     When starting a learner you must pass a function to be invoked whenever
